Implement Default trait for various structs, optimize code by replacing or_insert_with with or_default, and enhance error handling in multiple modules. Update grammar conversion logic and improve test assertions for better clarity. Add scripts for clippy warnings and sequential publish verification.

+++ tablegen/src/abi_builder.rs
@@ -1157,7 +1157,7 @@ impl<'a> AbiLanguageBuilder<'a> {
         let mut index_to_id_entries = vec![quote! { 0 }; symbol_count];
 
         for (symbol_id, &index) in &self.parse_table.symbol_to_index {
-            let symbol_id_val = symbol_id.0 as u16;
+            let symbol_id_val = symbol_id.0;
             index_to_id_entries[index] = quote! { #symbol_id_val };
         }
